Adding new property pages.
