+++ PowerShellTools/Guids.cs
@@ -8,11 +8,22 @@ namespace PowerShellTools
     static class GuidList
     {
         public const string PowerShellLanguage = "1C4711F1-3766-4F84-9516-43FA4169CC36";
+
+        //Packages
         public const string PowerShellToolsPackageGuid = "58dce676-42b0-4dd6-9ee4-afbc8e582b8a";
         public const string PowerShellToolsProjectPackageGuid = "2F99237E-E34F-4A3D-A337-500E4B3167B8";
-        public const string PowerShellGeneralPropertiesPageGuid = "C9619BDD-D1B3-4ACA-ADF3-2323EB62315E";
-        public const string PowerShellModulePropertiesPageGuid = "C9619BDD-D1B3-4ACA-ADF3-2323EB623154";
-        public const string PowerShellDebugPropertiesPageGuid = "E6C81F79-910C-4C91-B9DF-321883DC9F44";
+
+        //Property Pages
+        public const string GeneralPropertiesPageGuid = "C9619BDD-D1B3-4ACA-ADF3-2323EB62315E";
+        public const string ModulePropertiesPageGuid = "C9619BDD-D1B3-4ACA-ADF3-2323EB623154";
+        public const string DebugPropertiesPageGuid = "E6C81F79-910C-4C91-B9DF-321883DC9F44";
+
+        public const string RequirementsPropertiesPageGuid = "C83BB2CD-5AD0-4A67-A7EC-1F3AAA53BF40";
+        public const string ExportsPropertiesPageGuid = "0EFCF8AA-BD48-4D4B-B871-2D7B5C3374F7";
+        public const string InformationPropertiesPageGuid = "E4096CF0-A695-41B8-BEA7-1705729C633C";
+        public const string ComponentsPropertiesPageGuid = "3B19A81A-BE20-4F0D-B577-5093853FB681";
+
+        //Commands
         public const string CmdSetGuid = "099073C0-B561-4BC1-A847-92657B89A00E";
         public const uint CmdidExecuteAsScript =  0x0102;
         public const uint CmdidExecuteSelection = 0x0103;
