+++ PowerShellTools/PowerShellToolsPackage.cs
@@ -23,6 +23,7 @@ using PowerShellTools.Diagnostics;
 using PowerShellTools.LanguageService;
 using PowerShellTools.Project;
 using log4net;
+using PowerShellTools.Project.PropertyPages;
 using Engine = PowerShellTools.DebugEngine.Engine;
 
 namespace PowerShellTools
@@ -57,9 +58,11 @@ namespace PowerShellTools
     //         DefaultName = "CustomEditor")]
     [ProvideKeyBindingTable(GuidList.guidCustomEditorEditorFactoryString, 102)]
     [Guid(GuidList.PowerShellToolsPackageGuid)]
-    //[ProvideObject(typeof (PowerShellGeneralPropertyPage))]
-    [ProvideObject(typeof (PowerShellModulePropertyPage))]
-    [ProvideObject(typeof(PowerShellDebugPropertyPage))]
+    [ProvideObject(typeof (InformationPropertyPage))]
+    [ProvideObject(typeof(ComponentsPropertyPage))]
+    [ProvideObject(typeof(ExportsPropertyPage))]
+    [ProvideObject(typeof(RequirementsPropertyPage))]
+    [ProvideObject(typeof(DebugPropertyPage))]
     [Microsoft.VisualStudio.Shell.ProvideDebugEngine("{43ACAB74-8226-4920-B489-BFCF05372437}", "PowerShell",
         PortSupplier = "{708C1ECA-FF48-11D2-904F-00C04FA302A1}",
         ProgramProvider = "{08F3B557-C153-4F6C-8745-227439E55E79}", Attach = true,
