+++ PowerShellTools/Project/PropertyPages/ComponentsPropertyPageControl.resx
